Deal with text/plain content type from AWS
Because AWS SNS only allows you to specify a single endpoint for
a subscription, the same URL must be used to handle both confirmation
and message handling.

In order to correctly route the request to the right controller
for each of these scenarios we need to look at the type of the
message that AWS is giving us.

Unfortunately, AWS sends these messages as text/plain so we can't
rely on rails to parse the JSON and need to instead look at headers.