Adds support for SES message delivery via S3
SES supports embedding messages in the SNS notification or storing
the message contents in S3 for later retrieval.

Messages embedded in SNS notifications have a maximum size of 150kb
[source](https://docs.aws.amazon.com/ses/latest/DeveloperGuide/receiving-email-action-sns.html)

Messages embedded in SNS also do not support attachments.

In order to support large messages and attachments the mail message
can be stored in S3 and retrieved using the url embedded in the SNS
notification.